+++ lib/mix/tasks/weber.ex
@@ -8,6 +8,7 @@ defmodule Mix.Tasks.Weber do
 
       mix weber.new /home/user/myWebApp - Creates myWebApp directory with weber project skeleton.
       mix weber --version - Prints weber version.
+      mix weber --help - Prints help.
   """
 
   @shortdoc "Create a new weber project"
@@ -29,7 +30,7 @@ defmodule Mix.Tasks.Weber do
   end
 
   def usage do
-    """
+    Mix.shell.info """
     Usage:
 
       mix weber.new /home/user/testWebApp -- creates new weber web application
@@ -79,56 +80,13 @@ defmodule Mix.Tasks.Weber do
       end
     end
 
-    @doc """
-    Create weber based project with utils
-    """
-    def run([directoryName | rest]) do
-      run([directoryName])
-      Enum.each(rest, 
-        fn(addition) -> 
-          case addition do
-            "--grunt" -> add_grunt(directoryName)
-            _ -> :wrong
-          end
-        end)
-    end
-
-    defp add_grunt(directoryName) do
-      path = Path.absname directoryName
-      basename = Path.basename(directoryName)
-      Mix.Generator.create_file path <> <<"/package.json">>, (package_json basename)
-      Mix.Generator.create_file path <> <<"/Gruntfile.js">>, (gruntfile)
-    end
-
-    defp package_json(basename) do
-      """
-      {
-        "name": "#{basename}",
-        "version": "0.0.1",
-        "devDependencies": {
-          "grunt": "~0.4.2"
-        }
-      }
-      """
-    end
-
-    defp gruntfile do
-      """
-      module.exports = function(grunt) {
-        grunt.initConfig({
-          pkg: grunt.file.readJSON('package.json')
-        });
-      };
-      """
-    end
-
     defp replace(text, vars) do
       case Enum.uniq(Regex.scan(%r/\#\{([^\}]+)\}/, text)) do
         []   -> text
         data -> replace_act(text, vars, data)
       end
     end
-
+ 
     defp replace_act(text, _vars, []) do
       text
     end
@@ -136,7 +94,6 @@ defmodule Mix.Tasks.Weber do
     defp replace_act(text, vars, [[entry, key] | tail]) do
       replace_act(String.replace(text, entry, HashDict.get(vars, key), []), vars, tail)
     end
-
+    
   end
-
 end